added another component for the detail and cleaned the code for the pages
