+++ code/src/components/Error.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export const Error = () => {
+  return (
+    <>
+      <section>
+        <h1>Oh no, there's an error</h1>
+        <a className="return-button" href="/">Return</a>
+      </section>
+    </>
+  )
+}
